# Optimized EnumSet::count() and EnumSet::getOrdinals()